requirements.txt
cffi==1.17.1
cryptography==44.0.2
pycparser==2.22
wheel @ file:///opt/homebrew/Cellar/python%403.13/3.13.2/libexec/wheel-0.45.1-py3-none-any.whl#sha256=b9235939e2096903717cb6bfc132267f8a7e46deb2ec3ef9c5e234ea301795d0
Scrapy>=2.11.0
beautifulsoup4>=4.12.0
pymongo>=4.6.0
python-dotenv>=1.0.0
